Enemy now flings itself towards the player. Fling settings can be customised in the Enemy Data scriptable objects.

+++ GMTK2022/Assets/Scripts/Dye.cs
@@ -56,7 +56,7 @@ public class Dye : MonoBehaviour
         hasJustPerformedAbility = true;
     }
 
-    protected virtual IEnumerator DelayFaceAbility(int delayTime)
+    protected virtual IEnumerator DelayFaceAbility(float delayTime)
     {
         //Debug.Log($"{name}: Waiting to perform Face Ability.", this);
         waitingToPerformAbility = true;
